requirements.txt
azure-eventhub==5.11.*
influxdb-client==1.36.*
python-dotenv==1.0.*